+++ Dockerfile
@@ -3,7 +3,10 @@
 FROM golang:1.22 AS builder
 WORKDIR /go/src/github.com/OwO-Network/DeepLX
 COPY main.go ./
-COPY dto.go ./
+COPY types.go ./
+COPY utils.go ./
+COPY config.go ./
+COPY translate.go ./
 COPY go.mod ./
 COPY go.sum ./
 RUN go get -d -v ./
